# openapi.yaml
openapi: 3.0.3
info:
  title: mock-mobile-money-service
  version: 1.0.0
  description: |
    Faux fournisseur **MVola** et **Orange Money** pour le Payment Service NovaMarket. Les comptes et le ledger sont
    persistés dans PostgreSQL; l'état technique des tokens provider et des transactions provider reste en mémoire.

    L'API utilisateur et l'Admin API sont consommées depuis des frontends distincts configurés avec
    `USER_FRONTEND_ORIGINS` et `ADMIN_FRONTEND_ORIGINS`. En production, les deux variables sont obligatoires et ne
    peuvent pas contenir `*`. Les méthodes CORS sont GET, POST, PUT, PATCH, DELETE et OPTIONS; les headers autorisés
    sont Content-Type et Authorization. Les routes provider ne dépendent pas du CORS navigateur. CORS n'authentifie
    aucune requête.

    Légende des tags :
    - **MVola** / **Orange Money** : contrat reproduit depuis la documentation / les SDK publics (voir README pour les
      éléments marqués ASSUMPTION : non vérifiables publiquement).
    - **Mock** : endpoints MOCK-ONLY sous `/__mock`, absents des vrais fournisseurs.
servers:
  - url: http://localhost:4010
    description: Mock local (MVOLA_BASE_URL et ORANGE_MONEY_BASE_URL pointent ici)
...
paths:
  /auth/register:
    post:
      tags: [User]
      operationId: registerUser
      summary: Créer un wallet utilisateur
      description: Le serveur génère un MSISDN unique selon les préfixes backend du provider. Le client ne peut pas choisir le MSISDN ni le solde.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
            example: { provider: MVOLA, password: secret123 }
      responses:
        "201":
          description: Compte créé; le hash du mot de passe n'est jamais retourné.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountResponse" }
              example: { provider: MVOLA, msisdn: "0341234567", balance: 0, currency: MGA, status: ACTIVE }
        "400":
          description: Provider ou mot de passe invalide
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "503":
          description: Impossible d'allouer un MSISDN unique
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /auth/login:
    post:
      tags: [User]
      operationId: loginUser
      summary: Authentifier un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
            example: { msisdn: "0341234567", password: secret123 }
      responses:
        "200":
          description: Jeton d'accès utilisateur signé, distinct de MOCK_ADMIN_TOKEN
          content:
            application/json:
              schema:
                type: object
                required: [accessToken, tokenType, expiresIn]
                properties:
                  accessToken: { type: string }
                  tokenType: { type: string, enum: [Bearer] }
                  expiresIn: { type: integer, example: 3600 }
        "400":
          description: Requête invalide
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: MSISDN ou mot de passe incorrect
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Compte bloqué
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /account:
    get:
      tags: [User]
      operationId: getCurrentAccount
      summary: Consulter son propre compte
      description: Le compte est dérivé du sujet du token; aucun MSISDN de compte n'est accepté comme sélecteur.
      security:
        - userBearer: []
      responses:
        "200":
          description: Compte authentifié
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountResponse" }
        "401":
          description: Token utilisateur absent, invalide ou expiré
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Compte bloqué
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /account/balance:
    get:
      tags: [User]
      operationId: getCurrentBalance
      summary: Consulter son solde
      security:
        - userBearer: []
      responses:
        "200":
          description: Solde du compte authentifié
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BalanceResponse" }
        "401":
          description: Token utilisateur absent, invalide ou expiré
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Compte bloqué
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /account/transactions:
    get:
      tags: [User]
      operationId: getCurrentAccountTransactions
      summary: Lister les transactions de son compte
      security:
        - userBearer: []
      responses:
        "200":
          description: Transactions du seul compte authentifié
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountTransactionListResponse" }
        "401":
          description: Token utilisateur absent, invalide ou expiré
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Compte bloqué
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /account/topup:
    post:
      tags: [User]
      operationId: topUpCurrentAccount
      summary: Créditer son compte
      description: Les top-ups ne sont pas limités en nombre; chaque crédit atomique est refusé si le solde dépasserait MAX_BALANCE.
      security:
        - userBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TopUpRequest" }
            example: { amount: 500000 }
      responses:
        "200":
          description: Solde crédité et transaction TOPUP enregistrée
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountResponse" }
        "400":
          description: Montant invalide ou inférieur ou égal à zéro
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Token utilisateur absent, invalide ou expiré
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Compte bloqué
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "422":
          description: Le solde maximal serait dépassé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /account/reset:
    post:
      tags: [User]
      operationId: resetCurrentAccount
      summary: Remettre son propre solde à zéro
      security:
        - userBearer: []
      responses:
        "200":
          description: Solde remis à zéro; une transaction RESET est créée
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountResponse" }
        "401":
          description: Token utilisateur absent, invalide ou expiré
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "403":
          description: Compte bloqué
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/accounts:
    get:
      tags: [Admin]
      operationId: adminListAccounts
      summary: Lister les comptes
      security:
        - adminBearer: []
      responses:
        "200":
          description: Liste des comptes, sans hashes de mots de passe
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AccountResponse" }
        "401":
          description: Token admin absent ou invalide
  /admin/accounts/{msisdn}:
    get:
      tags: [Admin]
      operationId: adminGetAccount
      summary: Consulter un compte
      security:
        - adminBearer: []
      parameters:
        - name: msisdn
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Compte
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountResponse" }
        "401":
          description: Token admin absent ou invalide
        "404":
          description: Compte inconnu
  /admin/accounts/{msisdn}/transactions:
    get:
      tags: [Admin]
      operationId: adminListAccountTransactions
      summary: Lister les transactions d'un compte
      security:
        - adminBearer: []
      parameters:
        - name: msisdn
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Transactions du compte
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TransactionResponse" }
        "401":
          description: Token admin absent ou invalide
...
components:
  securitySchemes:
    userBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton utilisateur signé avec USER_AUTH_SECRET, distinct de MOCK_ADMIN_TOKEN.
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Valeur de `MOCK_ADMIN_TOKEN`.
    basicAuth:
      type: http
      scheme: basic
      description: consumerKey:consumerSecret (MVola) ou header d'autorisation de l'app Orange. Le mock accepte toute valeur non vide.
    bearerAuth:
      type: http
      scheme: bearer
      description: Token obtenu via /token (MVola) ou /oauth/v3/token (Orange). Chaque fournisseur a ses propres tokens.
...
  schemas:
    RegisterRequest:
      type: object
      required: [provider, password]
      additionalProperties: false
      properties:
        provider: { type: string, enum: [MVOLA, ORANGE_MONEY] }
        password: { type: string, minLength: 8, maxLength: 128, format: password }
    LoginRequest:
      type: object
      required: [msisdn, password]
      additionalProperties: false
      properties:
        msisdn: { type: string, pattern: "^\\d{10}$" }
        password: { type: string, minLength: 8, maxLength: 128, format: password }
    AccountResponse:
      type: object
      required: [provider, msisdn, balance, currency, status]
      properties:
        provider: { type: string, enum: [MVOLA, ORANGE_MONEY] }
        msisdn: { type: string, pattern: "^\\d{10}$" }
        balance: { type: number, minimum: 0 }
        currency: { type: string, example: MGA }
        status: { type: string, enum: [ACTIVE, BLOCKED] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BalanceResponse:
      type: object
      required: [balance, currency]
      properties:
        balance: { type: number, minimum: 0 }
        currency: { type: string, example: MGA }
    TopUpRequest:
      type: object
      required: [amount]
      additionalProperties: false
      properties:
        amount: { type: number, minimum: 0, exclusiveMinimum: true }
    TransactionResponse:
      type: object
      required: [id, accountId, provider, type, amount, currency, status, reference, balanceBefore, balanceAfter, createdAt]
      properties:
        id: { type: string, format: uuid }
        accountId: { type: string, format: uuid }
        provider: { type: string, enum: [MVOLA, ORANGE_MONEY] }
        type: { type: string, enum: [PAYMENT, TOPUP, RESET, REFUND] }
        amount: { type: number, minimum: 0 }
        currency: { type: string }
        status: { type: string, enum: [SUCCESS, FAILED] }
        reference: { type: string }
        balanceBefore: { type: number, minimum: 0 }
        balanceAfter: { type: number, minimum: 0 }
        createdAt: { type: string, format: date-time }
    AccountTransactionListResponse:
      type: object
      required: [transactions]
      properties:
        transactions:
          type: array
          items: { $ref: "#/components/schemas/TransactionResponse" }
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
    # ---- MVola
...